refactor: move `expect` parsing utils into `parseExpectCall` file
